refactor: split NotebookCheckRenderer.swift (1,017 lines → 4 files), per #491
Renderer for in-browser notebook checks; each check kind has its own
~80-line render+default-label pair.  Split along check-type boundaries
so each file is one cohesive group.

Layout:
  NotebookCheckRenderer.swift           — dispatcher, top utilities,
                                          spec hash, shared escape
                                          helpers (151)
  NotebookCheckRenderer+DataFrame.swift — .dataFrameShape, .dataFrameColumns,
                                          .dataFrameEquality, .seriesEquality (339)
  NotebookCheckRenderer+Plots.swift     — .numericArrayClose, .figureCount,
                                          .cellContains (224)
  NotebookCheckRenderer+Code.swift      — .functionExists, .variableExists,
                                          .astStructure (324)

Visibility: per-check render*/default*Label/tierFilenamePrefixForCheck/
escapeForPythonStringLiteralCheck promoted from `private` to internal so
the dispatcher in the main file (and shared escape helpers) reach across.

Tests pass.
